Refactor code directory
