+++ src/components/user/dashboard/activities/ActivityCard.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import ActivityList from './ActivityList';
 
 const ActivityCard = () => (
-  <div>
-    <div className="border border-1-solid rounded-[12px] p-5 w-[300px]">
+  <div className="absolute top-[100px] left-[250px] w-[82%] px-10">
+    <div className="border border-1-solid rounded-[12px] p-5">
       <div className="flex flex-row items-center justify-between">
         <h2 className="text-[#A03]">Recent Activities</h2>
         <p>Date</p>
