+++ src/components/user/dashboard/common/UserCards.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
 // import Footer from '../../common/Footer';
-import ActivityCard from './ActivityCard';
-import DonationCard from './DonationCard';
-import RequestCard from './RequestCard';
+// import ActivityCard from './ActivityCard';
+// import DonationCard from './DonationCard';
+// import RequestCard from './RequestCard';
 
 const UserCards = () => (
   <section className="flex flex-col lg:block items-center justify-center w-[100%] mt-[50px]">
     <div className="flex gap-5 flex-col lg:flex-row">
-      <DonationCard />
-      <RequestCard />
-      <ActivityCard />
+      {/* <DonationCard /> */}
+      {/* <RequestCard /> */}
+      {/* <ActivityCard /> */}
     </div>
   </section>
 );
